parallel index scan

+++ src/backend/gpopt/translate/CTranslatorQueryToDXL.cpp
@@ -3571,7 +3571,7 @@ CTranslatorQueryToDXL::TranslateValueScanRTEToDXL(const RangeTblEntry *rte,
 	const ULONG num_of_tuples = gpdb::ListLength(tuples_list);
 	GPOS_ASSERT(0 < num_of_tuples);
 
-	if (num_of_tuples > optimizer_valuescan_threshold)
+	if (num_of_tuples > static_cast<ULONG>(optimizer_valuescan_threshold))
 	{
 		GPOS_RAISE(gpdxl::ExmaDXL, gpdxl::ExmiQuery2DXLError,
 				   GPOS_WSZ_LIT("too many tuples in VALUES clause"));
